add filter on already computed mutations analysis

+++ src/main/python/mutation_analysis_on_amplification.py
@@ -2,6 +2,7 @@ import json
 import subprocess
 import sys
 from os import walk
+import os.path
 
 import profiling_test_class
 import build_rate_table
@@ -35,33 +36,35 @@ def run(projects, mvnHome="~/apache-maven-3.3.9/bin/"):
             test_name_ampl = '.'.join(element[2].split(".")[:-1]) + '.' + \
                              ("Ampl" + element[2].split(".")[-1] if element[2].split(".")[-1].endswith("Test") else
                               element[2].split(".")[-1] + "Ampl")
-            print test_name_ampl
-            filename = prefix + project + "/" + test_name_ampl.replace(".", "/") + ".java"
-            cmd = "cp " + filename + " " + path_to_test_directory
-            print cmd
-            #subprocess.call(cmd, shell=True)
 
-            cmd = "cd " + path_to_module + " && "
-            cmd += run_pitest
-            cmd += report
-            cmd += targetClasses + properties["filter"]
-            cmd += test_class + test_name_ampl
-            if "additionalClasspathElements" in properties:
-                cmd += " -DadditionalClasspathElements=" + properties["additionalClasspathElements"]
-            if "excludedClasses" in properties:
-                cmd += " -DexcludedClasses=" + properties["excludedClasses"]
-            print cmd
-            #subprocess.call(cmd, shell=True)
+            if not os.path.isfile(prefix + project + "/" + test_name_ampl + "_mutations.csv"):
+                print test_name_ampl
+                filename = prefix + project + "/" + test_name_ampl.replace(".", "/") + ".java"
+                cmd = "cp " + filename + " " + path_to_test_directory
+                print cmd
+                #subprocess.call(cmd, shell=True)
 
-            path_to_pit_reports = path_to_module + "/target/pitest-reports/"
-            for (dirpathpit, dirnamespit, filenamespit) in walk(path_to_pit_reports):
-                if filenamespit:
-                    for filenamepit in filenamespit:
-                        if filenamepit.endswith(".csv"):
-                            cmd = "cp " + dirpathpit + "/" + filenamepit + " " + \
-                                  prefix + project + "/" + test_name_ampl + "_mutations.csv"
-                            print cmd
-                            #subprocess.call(cmd, shell=True)
+                cmd = "cd " + path_to_module + " && "
+                cmd += run_pitest
+                cmd += report
+                cmd += targetClasses + properties["filter"]
+                cmd += test_class + test_name_ampl
+                if "additionalClasspathElements" in properties:
+                    cmd += " -DadditionalClasspathElements=" + properties["additionalClasspathElements"]
+                if "excludedClasses" in properties:
+                    cmd += " -DexcludedClasses=" + properties["excludedClasses"]
+                print cmd
+                #subprocess.call(cmd, shell=True)
+
+                path_to_pit_reports = path_to_module + "/target/pitest-reports/"
+                for (dirpathpit, dirnamespit, filenamespit) in walk(path_to_pit_reports):
+                    if filenamespit:
+                        for filenamepit in filenamespit:
+                            if filenamepit.endswith(".csv"):
+                                cmd = "cp " + dirpathpit + "/" + filenamepit + " " + \
+                                      prefix + project + "/" + test_name_ampl + "_mutations.csv"
+                                print cmd
+                                #subprocess.call(cmd, shell=True)
 
 
 if __name__ == '__main__':
